[edit] : SeatLayout 의 모달 내부 jsx를 컴포넌트로 분리
