+++ src/components/seats/SeatLayout.tsx
@@ -10,6 +10,8 @@ import { bookResultType } from "@/types/reservation";
 import { Timestamp } from "firebase/firestore";
 import { CloseButton } from "../controls/Button";
 import Loading from "../Loading";
+import BookedSeatInfo from "../booking/BookedSeatInfo";
+import NonBookedSeatInfo from "../booking/NonBookedSeatInfo";
 
 interface SeatLayoutProps {
   rows: number;
@@ -25,12 +27,6 @@ interface SeatLayoutProps {
 interface BookResult extends bookResultType {
   bookingTimestamp: Timestamp;
 }
-
-const BookingStatus = {
-  processing: "예매중",
-  pending: "입금 대기",
-  booked: "입금 완료",
-};
 
 export default function SeatLayout({
   rowsConfigs,
@@ -109,123 +105,6 @@ export default function SeatLayout({
     });
   };
 
-  const errorInfo = () => {
-    return (
-      <section className="border-t-2 py-2 text-flesh-500">
-        <h3 className="font-bold">Error</h3>
-        <p>{bookingError}</p>
-      </section>
-    );
-  };
-
-  const renderBookedSeatInfo = () => {
-    if (selectedSeatItem == undefined) {
-      return;
-    }
-
-    return (
-      <article>
-        <h2 className="text-2xl font-bold text-center">
-          좌석 {selectedSeatLabel} 정보
-        </h2>
-        <dl className="flex flex-wrap py-2">
-          <dt className="basis-[30%] text-gray-500">예매번호 :</dt>
-          <dd className="font-bold basis-[70%] break-all">
-            {selectedSeatItem[0].reservationId}
-          </dd>
-        </dl>
-        <section className="border-t-2 py-2">
-          <h3 className="font-bold">공연 정보</h3>
-          <dl className="flex flex-wrap items-center justify-center">
-            <dt className="basis-[30%] text-gray-500 my-0.5">공연명</dt>
-            <dd className="basis-[70%] my-0.5">{performanceData?.title}</dd>
-            <dt className="basis-[30%] text-gray-500 my-0.5">공연 날짜</dt>
-            <dd className="basis-[70%] my-0.5">{performanceDate}</dd>
-            <dt className="basis-[30%] text-gray-500 my-0.5">공연 시간</dt>
-            <dd className="basis-[70%] my-0.5">{performanceTime}</dd>
-            <dt className="basis-[30%] text-gray-500 my-0.5">가격</dt>
-            <dd className="basis-[70%] my-0.5">
-              {performanceData?.price.toLocaleString()} 원
-            </dd>
-          </dl>
-        </section>
-        {bookingError != null ? (
-          errorInfo()
-        ) : (
-          <>
-            <section className="border-t-2 py-2">
-              <h3 className="font-bold">예약 정보</h3>
-              <dl className="flex flex-wrap items-center justify-center">
-                <dt className="basis-[30%] text-gray-500 my-0.5">예약자명</dt>
-                <dd className="basis-[70%] my-0.5">
-                  {bookingData?.purchaserInfo.name}
-                </dd>
-                <dt className="basis-[30%] text-gray-500 my-0.5">이메일</dt>
-                <dd className="basis-[70%] my-0.5">
-                  {bookingData?.purchaserInfo.email}
-                </dd>
-                <dt className="basis-[30%] text-gray-500 my-0.5">연락처</dt>
-                <dd className="basis-[70%] my-0.5">
-                  {bookingData?.purchaserInfo.phone}
-                </dd>
-                <dt className="basis-[30%] text-gray-500 my-0.5">예약 상태</dt>
-                <dd className="basis-[70%] my-0.5">
-                  {bookingData?.paymentStatus &&
-                    BookingStatus[
-                      bookingData.paymentStatus as keyof typeof BookingStatus
-                    ]}
-                </dd>
-                <dt className="basis-[30%] text-gray-500 my-0.5">
-                  추가 예약 좌석
-                </dt>
-                <dd className="basis-[70%] my-0.5">
-                  {bookingData?.selectedSeats.filter(
-                    (seatlabel) => seatlabel !== selectedSeatLabel
-                  )}
-                </dd>
-              </dl>
-            </section>
-
-            <section className="border-t-2 py-2">
-              <h3 className="font-bold">결제 정보</h3>
-              <dl className="flex flex-wrap items-center justify-center">
-                <dt className="basis-[30%] my-0.5">총 결제 금액</dt>
-                <dd className="basis-[70%] my-0.5 text-right font-bold">
-                  {bookingData?.totalPrice.toLocaleString()} 원
-                </dd>
-              </dl>
-            </section>
-          </>
-        )}
-      </article>
-    );
-  };
-
-  const renderNonBookedSeatInfo = () => {
-    return (
-      <article>
-        <h2 className="text-2xl font-bold text-center mb-2">
-          좌석 {selectedSeatLabel} 정보
-        </h2>
-        <section className="border-t-2 py-2">
-          <h3 className="font-bold">공연 정보</h3>
-          <dl className="flex flex-wrap items-center justify-center">
-            <dt className="basis-[30%] text-gray-500 my-0.5">공연명</dt>
-            <dd className="basis-[70%] my-0.5">{performanceData?.title}</dd>
-            <dt className="basis-[30%] text-gray-500 my-0.5">공연 날짜</dt>
-            <dd className="basis-[70%] my-0.5">{performanceDate}</dd>
-            <dt className="basis-[30%] text-gray-500 my-0.5">공연 시간</dt>
-            <dd className="basis-[70%] my-0.5">{performanceTime}</dd>
-            <dt className="basis-[30%] text-gray-500 my-0.5">가격</dt>
-            <dd className="basis-[70%] my-0.5">
-              {performanceData?.price.toLocaleString()} 원
-            </dd>
-          </dl>
-        </section>
-      </article>
-    );
-  };
-
   const renderLoadingState = () => {
     return <Loading />;
   };
@@ -262,11 +141,28 @@ export default function SeatLayout({
           <CloseButton className="py-2 ml-auto" onClick={handleCloseModal} />
           {isBooked &&
           selectedSeatItem &&
-          selectedSeatItem[0].status !== "processing"
-            ? bookingLoading
-              ? renderLoadingState()
-              : renderBookedSeatInfo()
-            : renderNonBookedSeatInfo()}
+          selectedSeatItem[0].status !== "processing" ? (
+            bookingLoading ? (
+              renderLoadingState()
+            ) : (
+              <BookedSeatInfo
+                selectedSeatLabel={selectedSeatLabel}
+                selectedSeatItem={selectedSeatItem[0]}
+                performanceData={performanceData}
+                performanceDate={performanceDate}
+                performanceTime={performanceTime}
+                bookingData={bookingData}
+                bookingError={bookingError}
+              />
+            )
+          ) : (
+            <NonBookedSeatInfo
+              selectedSeatLabel={selectedSeatLabel}
+              performanceData={performanceData}
+              performanceDate={performanceDate}
+              performanceTime={performanceTime}
+            />
+          )}
         </>
       </Modal>
     </>
